move carousel to gatsby scope
