--- libs/react-components/src/lib/HeroCarousel/HeroCarousel.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from "@testing-library/react";
-
-import HeroCarousel from "./HeroCarousel";
-
-describe("HeroCarousel", () => {
-  it("should render successfully", () => {
-    //const { baseElement } = render(<HeroCarousel />);
-    //expect(baseElement).toBeTruthy();
-  });
-});
